[pig] the 'buildPodMemory' can be a double
Also add the initial framework to validate a build-config.yaml

If the 'buildPodMemory' is defined in the build-config as either a
negative number, or it cannot be parsed as a number, an error will be
printed about this failed validation and the CLI stops. To do so, we
implement the `Validate` interface.

This also sets up the framework to validate more of build-config.yaml in
the future (See #300)